+++ lib/platform_plus.dart
@@ -1,6 +1,5 @@
 library platform_plus;
 
 export 'src/platform_plus_base.dart';
-export 'src/platform_plus_base.dart';
 export 'src/model/android_version_code.dart';
 export 'src/model/ios_device.dart';