+++ test/platform_test.dart
@@ -2,7 +2,8 @@ import 'package:flutter_test/flutter_test.dart';
 import 'package:platform_plus/platform_plus.dart';
 
 void main() {
-  test('isUnitTest', () {
-    expect(platformPlus.isUnitTest, true);
+  test('isUnitTest', () async {
+    final platform = await PlatformPlus.platform;
+    expect(platform.isUnitTest, true);
   });
 }